+++ docs/changes.rst
@@ -9,6 +9,8 @@ To be released.
 - Geofront now supports ECDSA key.
   Under the hood, :func:`parse_openssh_pubkey()` function becomes to be
   possible to parse ECDSA public keys.
+- Added :mod:`geofront.backends.oauth` module which provides
+  base team implementations for OAuth.
 - Now master key can be found without token through
   :http:get:`/masterkey/` API.
 - Deprecated :http:get:`/tokens/(token_id:token_id)/masterkey/`.